Move autoBody render to transformation
